+++ talos-cluster/providers.tf
@@ -1,10 +1,5 @@
 terraform {
   required_providers {
-    helm = {
-      source  = "hashicorp/helm"
-      version = "~> 3.1"
-    }
-
     null = {
       source  = "hashicorp/null"
       version = "~> 3.2"