+++ CHANGELOG.md
@@ -8,7 +8,8 @@ No unreleased changes
 
 * Fixed README for `talos-cluster` module as it had an error with the output fields
 * Fixed the `mac_address` local variable for `talos-vm` as it was missing an index
-  
+* Fixed an incorrect reference to `ip_address` in the `talos-vm` module
+
 ## v0.1.0 (Released 2025-12-14)
 
 * Initial release of the module